flights/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
Flask-Cors==5.0.0
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-exporter-otlp==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-system-metrics
